requirements.txt
anyio==4.9.0
asgiref==3.8.1
beautifulsoup4==4.13.4
blinker==1.9.0
certifi==2025.4.26
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.1
colorama==0.4.6
Deprecated==1.2.18
dj-database-url==3.0.1
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-jazzmin==3.0.1
django-rest-framework==0.1.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Flask==3.1.1
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
lxml==5.4.0
MarkupSafe==3.0.2
MouseInfo==0.1.3
msgpack==1.1.1
packaging==25.0
phonenumbers==9.0.5
pikepdf==9.7.0
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
PyAutoGUI==0.9.54
pygame==2.6.1
PyGetWindow==0.0.9
PyJWT==2.9.0
PyMsgBox==1.0.9
pyperclip==1.9.0
PyRect==0.2.0
PyScreeze==1.0.1
python-decouple==3.8
pytweening==1.2.0
pywhatkit==5.4
questionary==2.1.0
redis==6.2.0
requests==2.32.3
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
sqlparse==0.5.3
truecallerpy==1.0.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
wcwidth==0.2.13
Werkzeug==3.1.3
whitenoise==6.9.0
wikipedia==1.4.0
wrapt==1.17.2
